+++ p2p/src/authenticated/discovery/actors/tracker/set.rs
@@ -22,7 +22,7 @@ impl<P: PublicKey> Set<P> {
         for (i, peer) in peers.iter().enumerate() {
             order.insert(peer.clone(), i);
         }
-        let knowledge = BitMap::zeroes(peers.len() as u64);
+        let knowledge = BitMap::zeroes(peers.len());
         Self {
             sorted: peers,
             order,
@@ -33,7 +33,7 @@ impl<P: PublicKey> Set<P> {
     /// Marks the given peer as known or unknown.
     pub fn update(&mut self, peer: &P, known: bool) -> bool {
         if let Some(idx) = self.order.get(peer) {
-            self.knowledge.set(*idx as u64, known);
+            self.knowledge.set(*idx, known);
             return true;
         }
         false
