+++ storage/src/adb/any/fixed/mod.rs
@@ -259,7 +259,7 @@ impl<
         mut bitmap: Option<&mut MerkleizedBitMap<H, N>>,
     ) -> Result<(), Error> {
         if let Some(ref bitmap) = bitmap {
-            assert_eq!(inactivity_floor_loc, bitmap.len());
+            assert_eq!(inactivity_floor_loc, bitmap.len() as u64);
         }
 
         let stream = log
@@ -279,7 +279,7 @@ impl<
                             if let Some(ref mut bitmap) = bitmap {
                                 // Mark previous location (if any) of the deleted key as inactive.
                                 if let Some(old_loc) = result {
-                                    bitmap.set_bit(old_loc, false);
+                                    bitmap.set_bit(old_loc as usize, false);
                                 }
                             }
                         }
@@ -288,7 +288,7 @@ impl<
                                 Any::<E, K, V, H, T>::update_loc(snapshot, log, &key, i).await?;
                             if let Some(ref mut bitmap) = bitmap {
                                 if let Some(old_loc) = result {
-                                    bitmap.set_bit(old_loc, false);
+                                    bitmap.set_bit(old_loc as usize, false);
                                 }
                                 bitmap.push(true);
                             }
@@ -298,7 +298,7 @@ impl<
                     if let Some(ref mut bitmap) = bitmap {
                         // If we reach this point and a bit hasn't been added for the operation, then it's
                         // an inactive operation and we need to tag it as such in the bitmap.
-                        if bitmap.len() == i {
+                        if bitmap.len() as u64 == i {
                             bitmap.push(false);
                         }
                     }
@@ -1384,7 +1384,7 @@ pub(super) mod test {
 
             // Check the bitmap state matches that of the snapshot.
             let items = db.log.size().await.unwrap();
-            assert_eq!(bitmap.len(), items);
+            assert_eq!(bitmap.len() as u64, items);
             let mut active_positions = HashSet::new();
             // This loop checks that the expected true bits are true in the bitmap.
             for pos in db.inactivity_floor_loc..items {
@@ -1398,7 +1398,7 @@ pub(super) mod test {
                     if *loc == pos {
                         // Found an active op.
                         active_positions.insert(pos);
-                        assert!(bitmap.get_bit(pos));
+                        assert!(bitmap.get_bit(pos as usize));
                         break;
                     }
                 }
@@ -1406,7 +1406,7 @@ pub(super) mod test {
             // This loop checks that the expected false bits are false in the bitmap.
             for pos in db.inactivity_floor_loc..items {
                 if !active_positions.contains(&pos) {
-                    assert!(!bitmap.get_bit(pos));
+                    assert!(!bitmap.get_bit(pos as usize));
                 }
             }
